shopflow-backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-filter==25.1
Markdown==3.8
pymongo==4.6.2
python-dotenv==1.0.1
psycopg==3.2.9
django-allauth==0.54.0
dj-rest-auth==3.0.0
djangorestframework_simplejwt==5.5.0
Pygments==2.19.1
PyJWT==2.9.0
sqlparse==0.5.3
google-auth requests